+++ src/sys.rs
@@ -1,10 +1,10 @@
 #[cfg(unix)]
-pub use self::unix::{exit, get_terminal_size};
+pub(crate) use self::unix::{exit, get_terminal_size};
 #[cfg(windows)]
-pub use self::winapi::{exit, get_terminal_size};
+pub(crate) use self::winapi::{exit, get_terminal_size};
 
 #[cfg(windows)]
-pub mod winapi;
+pub(crate) mod winapi;
 
 #[cfg(unix)]
-pub mod unix;
+pub(crate) mod unix;
