+++ src/sys/winapi.rs
@@ -1,13 +1,11 @@
 use crossterm_utils::Result;
 use crossterm_winapi::ScreenBuffer;
 
-/// Exit the current process.
-pub fn exit() {
+pub(crate) fn exit() {
     ::std::process::exit(256);
 }
 
-#[cfg(windows)]
-pub fn get_terminal_size() -> Result<(u16, u16)> {
+pub(crate) fn get_terminal_size() -> Result<(u16, u16)> {
     let terminal_size = ScreenBuffer::current()?.info()?.terminal_size();
     // windows starts counting at 0, unix at 1, add one to replicated unix behaviour.
     Ok((
