+++ src/terminal.rs
@@ -1,34 +1,14 @@
 //! A module that contains all the actions related to the terminal. like clearing, resizing, pausing and scrolling the terminal.
 use crossterm_utils::Result;
-#[cfg(feature = "serde")]
-use serde::{Deserialize, Serialize};
-
-use self::ansi::AnsiTerminal;
-pub use self::terminal::{Clear, ScrollDown, ScrollUp, SetSize, terminal, Terminal};
 #[cfg(windows)]
-use self::winapi::WinApiTerminal;
+use crossterm_utils::supports_ansi;
 
-mod terminal;
+use super::ClearType;
 
-mod ansi;
+pub(crate) mod ansi;
 #[cfg(windows)]
-mod winapi;
+pub(crate) mod winapi;
 
-/// Enum with the different values to clear the terminal.
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Ord, PartialOrd, Hash)]
-pub enum ClearType {
-    /// clear all cells in terminal.
-    All,
-    /// clear all cells from the cursor position downwards in terminal.
-    FromCursorDown,
-    /// clear all cells from the cursor position upwards in terminal.
-    FromCursorUp,
-    /// clear current line cells in terminal.
-    CurrentLine,
-    /// clear all cells from cursor position until new line in terminal.
-    UntilNewLine,
-}
 
 /// This trait defines the actions that can be performed with the terminal color.
 /// This trait can be implemented so that an concrete implementation of the ITerminalColor can fulfill.
@@ -38,7 +18,7 @@ pub enum ClearType {
 ///
 /// This trait is implemented for `WinApi` (Windows specific) and `ANSI` (Unix specific),
 /// so that terminal related actions can be performed on both Unix and Windows systems.
-trait ITerminal {
+pub(crate) trait Terminal {
     /// Clear the current cursor by specifying the clear type
     fn clear(&self, clear_type: ClearType) -> Result<()>;
     /// Get the terminal size (x,y)
